requirements.txt
python-telegram-bot>=20.0,<21.0
supabase>=2.0.0
python-dotenv>=1.0.0
jdatetime>=4.1.0

